Working undo state for select/transform

+++ src/editor/viewport/transform_tools/tool_state.go
@@ -1,6 +1,6 @@
 package transform_tools
 
-type ToolState = int
+type ToolState = uint8
 
 const (
 	ToolStateNone ToolState = iota
